Use upstream lit for all IREE lit testing
This removes IREE's hacky custom shell script, which seemed like a good
idea at the time.

I didn't bother trying to update the CMake from this side, but I'm
planning to make this same change in OSS, so we can copy back whatever
is needed from there, if desired. Given that, I just deleted the content
of the file to make it obvious how to resolve some future merge conflict.

One thing that needed to be fixed was changing our dependency on lld to
point explicitly at ld.lld, the expected name under unix. The
UnixLinkerTool looks for this basename and we were basically just
getting lucky that it happened to be able to find it. That makes the
Bazel config for these tests unix-specific, which given that they've
never really worked on any other platform, seems ok. If that's a
problem we can do some tricks with select to get the correct binary
name on each platform.

PiperOrigin-RevId: 422647247